Fix warning about int->float conversion loss of precision, will not actually happen as all input values can be represented

+++ stb_image.h
@@ -7126,7 +7126,7 @@ static void stbi__hdr_convert(float *output, stbi_uc *input, int req_comp)
    if ( input[3] != 0 ) {
       float f1;
       // Exponent
-      f1 = SDL_powf(2, input[3] - (int)(128 + 8));
+      f1 = SDL_powf(2, (float)((int)input[3] - (int)(128 + 8)));
       if (req_comp <= 2)
          output[0] = (input[0] + input[1] + input[2]) * f1 / 3;
       else {
